converts survivor creation to new request structure
